% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_projection.R
\name{bootstrap_projection}
\alias{bootstrap_projection}
\title{Bootstrap projection for network replication analysis}
\usage{
bootstrap_projection(B = 1000, M, X1, Y1, X2, Y2, npc = 3, alpha = 0.05)
}
\arguments{
\item{B}{number of bootstrap samples}

\item{M}{adjacency matrix}

\item{X1}{abundance data (group 1)}

\item{Y1}{phenotype data (group 1)}

\item{X2}{abundance data (group 2)}

\item{Y2}{phenotype data (group 2)}

\item{npc}{number of principal components}

\item{alpha}{significance level}
}
\value{
list containing correlations and bootstrap confidence intervals
}
\description{
Bootstrap projection for network replication analysis
}
