% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netshy_score.R
\name{netshy_score}
\alias{netshy_score}
\title{Compute NetSHY Summary Scores}
\usage{
netshy_score(A, X, npc = 1)
}
\arguments{
\item{A}{Symmetrical, non-negative network adjacency matrix.}

\item{X}{Numeric matrix of feature abundances (samples x features).}

\item{npc}{Integer. Number of principal components to extract.}
}
\value{
A matrix of samples by \code{npc} containing the network summary scores.
}
\description{
Compute NetSHY Summary Scores
}
